refactor(Server): 토큰 변경
