+++ app/src/main/java/woowacourse/shopping/data/model/entity/RecentProductEntity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.model
+package woowacourse.shopping.data.model.entity
 
 data class RecentProductEntity(
     val productId: Long,
